TypeScript module definitions
