+++ preact.d.ts
@@ -1,0 +1,19 @@
+// T - Wrapped component props
+// S - Wrapped component state
+// K - Store state
+// I - Injected props to wrapped component
+import * as Preact from "preact";
+import { ActionCreator, StateMapper, Store } from ".";
+
+export function connect<T, S, K, I>(
+  mapStateToProps: string | Array<string> | StateMapper<T, K, I>,
+  actions?: ActionCreator<K> | object
+): (Child: (props: T & I) => Preact.VNode) => Preact.ComponentConstructor<T, S>;
+
+export interface ProviderProps<T> {
+  store: Store<T>;
+}
+
+export class Provider<T> extends Preact.Component<ProviderProps<T>, {}> {
+  render(props: ProviderProps<T>, {}): Preact.VNode;
+}
